LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-anchor-fragment-expected.txt


FAIL Fragment <a> click fires navigate event assert_equals: expected (object) Element node <a id="a" href="#1"></a> but got (undefined) undefined
